4.2 commit

+++ src/main/java/com/vastika/Team_A_Account/service/AccountInfoService.java
@@ -7,8 +7,8 @@ import com.vastika.Team_A_Account.model.AccountInfo;
 public interface AccountInfoService {
 	int saveCustomerInfo(AccountInfo accInfo);
 	int updateCustomerInfo(AccountInfo accInfo);
-	void deleteCustomerInfo(long customerId);
-	AccountInfo getCustomerById (long customerId);
+	void deleteCustomerInfo(int customerId);
+	AccountInfo getCustomerById (int customerId);
 	List<AccountInfo> getAllCustomerList();
 	
 }
